builds with leafIndex as cpp_int

+++ src/matchgpu.h
@@ -35,7 +35,7 @@ class GraphMatchingGPU : public GraphMatching
 	public:
 		GraphMatchingGPU(const Graph &, const int &, const unsigned int &);
 		virtual ~GraphMatchingGPU();
-		virtual void performMatching(int *, cudaEvent_t &, cudaEvent_t &, int numberOfKernelCols, int * deviceKernelColumns, int numberOfTreeVertsCols, int * deviceTreeColumns, int numberOfDynamicCols, int * deviceDynamicColumns, int leafIndex) const = 0;
+		virtual void performMatching(int *, cudaEvent_t &, cudaEvent_t &, int numberOfKernelCols, int * deviceKernelColumns, int numberOfTreeVertsCols, int * deviceTreeColumns, int numberOfDynamicCols, int * deviceDynamicColumns, cpp_int leafIndex) const = 0;
 		int2 *dneighbourRanges;
 		int *dneighbours;
 	protected:
@@ -49,7 +49,7 @@ class GraphMatchingGPURandom : public GraphMatchingGPU
 		GraphMatchingGPURandom(const Graph &, const int &, const unsigned int &);
 		~GraphMatchingGPURandom();
 		
-		void performMatching(int *, cudaEvent_t &, cudaEvent_t &, int numberOfKernelCols, int * deviceKernelColumns, int numberOfTreeVertsCols, int * deviceTreeColumns, int numberOfDynamicCols, int * deviceDynamicColumns, int leafIndex) const;
+		void performMatching(int *, cudaEvent_t &, cudaEvent_t &, int numberOfKernelCols, int * deviceKernelColumns, int numberOfTreeVertsCols, int * deviceTreeColumns, int numberOfDynamicCols, int * deviceDynamicColumns, cpp_int leafIndex) const;
 
 };
 
@@ -59,7 +59,7 @@ class GraphMatchingGeneralGPURandom : public GraphMatchingGPU
 		GraphMatchingGeneralGPURandom(const Graph &, const int &, const unsigned int &);
 		~GraphMatchingGeneralGPURandom();
 
-		void performMatching(int *, cudaEvent_t &, cudaEvent_t &, int numberOfKernelCols, int * deviceKernelColumns, int numberOfTreeVertsCols, int * deviceTreeColumns, int numberOfDynamicCols, int * deviceDynamicColumns, int leafIndex) const;
+		void performMatching(int *, cudaEvent_t &, cudaEvent_t &, int numberOfKernelCols, int * deviceKernelColumns, int numberOfTreeVertsCols, int * deviceTreeColumns, int numberOfDynamicCols, int * deviceDynamicColumns, cpp_int leafIndex) const;
 		int *drequests, *dsense;
 		int *dlength, *dforwardlinkedlist, *dbackwardlinkedlist, *dmatch, *ddegrees;
 		// Never directly used, just used for thrust::sequence functionality to avoid a kernel call.
@@ -74,7 +74,7 @@ class GraphMatchingGPURandomMaximal : public GraphMatchingGPU
 		GraphMatchingGPURandomMaximal(const Graph &, const int &, const unsigned int &);
 		~GraphMatchingGPURandomMaximal();
 		
-		void performMatching(int *, cudaEvent_t &, cudaEvent_t &, int numberOfKernelCols, int * deviceKernelColumns, int numberOfTreeVertsCols, int * deviceTreeColumns, int numberOfDynamicCols, int * deviceDynamicColumns, int leafIndex) const;
+		void performMatching(int *, cudaEvent_t &, cudaEvent_t &, int numberOfKernelCols, int * deviceKernelColumns, int numberOfTreeVertsCols, int * deviceTreeColumns, int numberOfDynamicCols, int * deviceDynamicColumns, cpp_int leafIndex) const;
 };
 
 class GraphMatchingGPUWeighted : public GraphMatchingGPU
@@ -83,7 +83,7 @@ class GraphMatchingGPUWeighted : public GraphMatchingGPU
 		GraphMatchingGPUWeighted(const Graph &, const int &, const unsigned int &);
 		~GraphMatchingGPUWeighted();
 		
-		void performMatching(int *, cudaEvent_t &, cudaEvent_t &, int numberOfKernelCols, int * deviceKernelColumns, int numberOfTreeVertsCols, int * deviceTreeColumns, int numberOfDynamicCols, int * deviceDynamicColumns, int leafIndex) const;
+		void performMatching(int *, cudaEvent_t &, cudaEvent_t &, int numberOfKernelCols, int * deviceKernelColumns, int numberOfTreeVertsCols, int * deviceTreeColumns, int numberOfDynamicCols, int * deviceDynamicColumns, cpp_int leafIndex) const;
 
 	private:
 		int *dweights;
@@ -95,7 +95,7 @@ class GraphMatchingGPUWeightedMaximal : public GraphMatchingGPU
 		GraphMatchingGPUWeightedMaximal(const Graph &, const int &, const unsigned int &);
 		~GraphMatchingGPUWeightedMaximal();
 		
-		void performMatching(int *, cudaEvent_t &, cudaEvent_t &, int numberOfKernelCols, int * deviceKernelColumns, int numberOfTreeVertsCols, int * deviceTreeColumns, int numberOfDynamicCols, int * deviceDynamicColumns, int leafIndex) const;
+		void performMatching(int *, cudaEvent_t &, cudaEvent_t &, int numberOfKernelCols, int * deviceKernelColumns, int numberOfTreeVertsCols, int * deviceTreeColumns, int numberOfDynamicCols, int * deviceDynamicColumns, cpp_int leafIndex) const;
 
 	private:
 		int *dweights;
